dataset_generator/requirements.txt
google-genai
python-dotenv
tqdm
streamlit
